Fix stats collector spinning CPU if no stats are collected
Commit 7ffb50b59375a6171795d5e59ec72059da2a0086 changed
the stats collection loop to use a `sleep()` instead
of `time.Tick()` in the for-loop.

This change caused a regression in situations where
no stats are being collected, or an error is hit
in the loop (in which case the loop would `continue`,
and the `sleep()` is not hit).

This patch puts the sleep at the start of the loop
to guarantee it's always hit.

This will delay the sampling, which is similar to the
behavior before 7ffb50b59375a6171795d5e59ec72059da2a0086.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>