Don't sort plugin mounts slice
This was added as part of 6a8fbf51210f103fa555e05d80fc096c5dfb7e00 with
the intent to sort the mounts in the plugin config, but this was sorting
*all* the mounts from the default OCI spec which is problematic.

In reality we don't need to sort this because we are only adding a
self-binded mount to flag it as rshared.

We may want to look at sorting the plugin mounts before they are added
to the OCI spec in the future, but for now I think the existing behavior
is fine since the plugin author has control of the order (except for the
propagated mount).

Signed-off-by: Brian Goff <cpuguy83@gmail.com>
(cherry picked from commit 0b48b65f435800278cc69e98e09a476280c2776c)
Signed-off-by: Brian Goff <cpuguy83@gmail.com>