aufs: remove extra locking
Both mount and unmount calls are already protected by fine-grained
(per id) locks in Get()/Put() introduced in commit 100760de9b4
("Add more locking to storage drivers"), so there's no point in
having a global lock in mount/unmount.

The only place from which unmount is called without any locking
is Cleanup() -- this is to be addressed in the next patch.

This reverts commit daf92f02b1c2f7c9298f5a5a08c90f452f893241.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>