Improve select for daemon restart tests
This improves the select logic for the restart tests or starting the
daemon in general.  With the way the ticker and select was setup, it was
possible for only the timeout to be displayed and not the wait errors.

Signed-off-by: Michael Crosby <crosbymichael@gmail.com>
(cherry picked from commit e97039db5b861eebeef8056e24827730d39562c6)
Signed-off-by: Sebastiaan van Stijn <github@gone.nl>